+++ src/flows/MessageFlow.js
@@ -41,11 +41,14 @@ Object.defineProperty(MessageFlow.prototype, 'counters', {
 });
 
 MessageFlow.prototype.getState = function getState() {
+  const brokerState = this.broker.getState(true);
+  if (!brokerState && this.environment.settings.disableTrackState) return;
+
   return {
     id: this.id,
     type: this.type,
     counters: this.counters,
-    broker: this.broker.getState(true),
+    broker: brokerState,
   };
 };