+++ src/flows/SequenceFlow.js
@@ -74,11 +74,14 @@ SequenceFlow.prototype.discard = function discard(content = {}) {
 };
 
 SequenceFlow.prototype.getState = function getState() {
+  const brokerState = this.broker.getState(true);
+  if (!brokerState && this.environment.settings.disableTrackState) return;
+
   return {
     id: this.id,
     type: this.type,
     counters: this.counters,
-    broker: this.broker.getState(true),
+    broker: brokerState,
   };
 };