+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+# 13.1.0
+
+- introduce `disableTrackState` setting. Tracking of elements is done by counters, e.g. activity taken or discarded, sequence flow taken and discarded. Counters are saved when getting state. If you run really big flows the state will keep all elements just to be able to recover the number of times an element has been touched. Needless to say it the state will grow out of it's comfort zone. Setting `disableTrackState` to true will only return state for elements that are actually running
+
+## Breaking
+
+- `getState()` can return undefined
+
 # 13.0.0
 
 - export task-, events-, and gateway activity behaviors through `bpmn-elements/tasks`, `bpmn-elements/events`, and `bpmn-elements/gateways` respectively